ssh: refactor libssh2-specific bits into their own file
