+++ src/libgit2/libgit2.c
@@ -34,7 +34,7 @@
 #include "streams/socket.h"
 #include "transports/smart.h"
 #include "transports/http.h"
-#include "transports/ssh.h"
+#include "transports/ssh_libssh2.h"
 
 #ifdef GIT_WIN32
 # include "win32/w32_leakcheck.h"
@@ -80,7 +80,7 @@ int git_libgit2_init(void)
 		git_sysdir_global_init,
 		git_filter_global_init,
 		git_merge_driver_global_init,
-		git_transport_ssh_global_init,
+		git_transport_ssh_libssh2_global_init,
 		git_stream_registry_global_init,
 		git_socket_stream_global_init,
 		git_openssl_stream_global_init,
