--- src/libgit2/transports/ssh.h
@@ -1,14 +1,0 @@
-/*
- * Copyright (C) the libgit2 contributors. All rights reserved.
- *
- * This file is part of libgit2, distributed under the GNU GPL v2 with
- * a Linking Exception. For full terms see the included COPYING file.
- */
-#ifndef INCLUDE_transports_ssh_h__
-#define INCLUDE_transports_ssh_h__
-
-#include "common.h"
-
-int git_transport_ssh_global_init(void);
-
-#endif
